+++ src/aux-server/aux-web/aux-projector/BuilderGameView/BuilderGameView.ts
@@ -463,7 +463,7 @@ export default class BuilderGameView extends Vue implements IGameView {
                     let newContext = contextMap.get(c);
                     worksurface.tags[c] = null;
                     worksurface.tags['aux.context'] = newContext;
-                    worksurface.tags['aux.context.surface'] = true;
+                    worksurface.tags['aux.context.visualize'] = 'surface';
                     worksurface.tags[newContext] = true;
                 });